Implement 'clear' command
Used to clear breakpoints.  Like the 'breakpoint' command, this is a bit
inflexible at the moment and requires a full filename and line number
pair to indicate the breakpoint to clear.

+++ src/ssj/source.h
@@ -5,7 +5,7 @@ typedef struct source source_t;
 
 source_t*        load_source     (const char* filename, const path_t* source_path);
 void             free_source     (source_t* source);
-const lstring_t* get_source_line (const source_t* source, size_t index);
-size_t           get_source_size (const source_t* source);
+const lstring_t* get_source_line (const source_t* source, int index);
+int              get_source_size (const source_t* source);
 
 #endif // SSJ__SOURCE_H__INCLUDED
